requirements.txt
streamlit
torch
torchvision
segmentation-models-pytorch
opencv-python-headless
numpy
Pillow
matplotlib
reportlab
scikit-image
pydicom
fpdf